# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Café Run Club Rewards API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /members/me:
    get:
      tags: [members]
      operationId: getMyProfile
      summary: Get my member profile
      responses:
        '200':
          description: Member profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags: [members]
      operationId: updateMyProfile
      summary: Update my display name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberUpdate'
      responses:
        '200':
          description: Updated member profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /members/me/dependents:
    get:
      tags: [members]
      operationId: listMyDependents
      summary: List my dependent profiles (kids, etc.)
      responses:
        '200':
          description: List of dependents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Member'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    post:
      tags: [members]
      operationId: addDependent
      summary: Add a dependent profile under my account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DependentInput'
      responses:
        '201':
          description: Created dependent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /members/me/dependents/{dependentId}:
    delete:
      tags: [members]
      operationId: removeDependent
      summary: Remove a dependent profile
      parameters:
        - name: dependentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Dependent removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteSuccess'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /members:
    get:
      tags: [members]
      operationId: listMembers
      summary: List all members (admin only)
      responses:
        '200':
          description: List of members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Member'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /admin/members/{userId}:
    get:
      tags: [members]
      operationId: getAdminMemberDetail
      summary: Get full member detail (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Member detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MemberDetail'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Member not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
    patch:
      tags: [members]
      operationId: adminUpdateMember
      summary: Update a member's profile (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemberUpdate'
      responses:
        '200':
          description: Updated member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Member'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    DeleteSuccess:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    Member:
      type: object
      required: [id, userId, displayName, totalMiles, joinedAt, isAdmin]
      properties:
        id:
          type: integer
        userId:
          type: string
        displayName:
          type: string
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        emergencyContact:
          type: ["string", "null"]
        emergencyPhone:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
        totalMiles:
          type: number
        joinedAt:
          type: string
          format: date-time
        isAdmin:
          type: boolean
        parentMemberId:
          type: ["integer", "null"]
    DependentInput:
      type: object
      required: [displayName]
      properties:
        displayName:
          type: string
          minLength: 1
    MemberUpdate:
      type: object
      properties:
        displayName:
          type: string
          minLength: 1
        email:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        emergencyContact:
          type: ["string", "null"]
        emergencyPhone:
          type: ["string", "null"]
    Run:
      type: object
      required: [id, userId, distanceMiles, date, createdAt, source]
      properties:
        id:
          type: integer
        userId:
          type: string
        distanceMiles:
          type: number
        date:
          type: string
          format: date
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        source:
          type: string
          enum: [manual, strava]
          default: manual
        stravaActivityId:
          type: ["string", "null"]
        clubEventId:
          type: ["integer", "null"]
...
    RewardTier:
      type: object
      required: [id, name, description, milesRequired, rewardType, active]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        milesRequired:
          type: number
        rewardType:
          type: string
          enum: [coffee, smoothie, apparel, custom]
        active:
          type: boolean
...
    MemberDetail:
      type: object
      required: [member, runs, earnedTiers]
      properties:
        member:
          $ref: '#/components/schemas/Member'
        runs:
          type: array
          items:
            $ref: '#/components/schemas/Run'
        earnedTiers:
          type: array
          items:
            $ref: '#/components/schemas/RewardTier'
        nextTier:
          oneOf:
            - $ref: '#/components/schemas/RewardTier'
            - type: 'null'
        milesUntilNext:
          type: ["number", "null"]
...